Make functions correspond more closely to Data.Map behavior
This is a breaking change. The more "monoidal" behaviors that this
reverts could eventuallly be re-added as a separate module that more
thoroughly and systematically applies the monoid/semigroup operations.

The goal of Data.Map.Monoidal is to be as similar to Data.Map as
possible except for the Monoid/Semigroup instance that it replaces.